% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcs.R
\name{longest_common_seq}
\alias{longest_common_seq}
\title{Find longest common subsequence between two strings.}
\usage{
longest_common_seq(x, y)
}
\arguments{
\item{x}{character string}

\item{y}{character string}
}
\value{
A longestCommonSeq object.
}
\description{
Find longest common subsequence between two strings.
}
\examples{

longest_common_seq(
  "Lorem dolor sit amet, consectetur adipiscing elit.",
  "Lorem ipsum dolor sit amet, eu erat sed felis pharetra.")

}
