% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcs.R
\name{print.longestCommonSeq}
\alias{print.longestCommonSeq}
\title{Print longestCommonSeq object}
\usage{
\method{print}{longestCommonSeq}(x)
}
\arguments{
\item{x}{longestCommonSeq pbject}
}
\description{
Print longestCommonSeq object
}
\examples{

lcs <- longest_common_seq(
  "Lorem dolor sit amet, consectetur adipiscing elit.",
  "Lorem ipsum dolor sit amet, eu erat sed felis pharetra.")

print(lcs)

}
