% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcs.R
\name{run_lcs_app}
\alias{run_lcs_app}
\title{Start the app for exploring longest common subsequence between two strings.}
\usage{
run_lcs_app(...)
}
\arguments{
\item{...}{arguments passed to the shiny's runApp function.}
}
\description{
Start the app for exploring longest common subsequence between two strings.
}
